update gen code structure
