+++ backend/app/generator/crud/crud_gen_model.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+from sqlalchemy_crud_plus import CRUDPlus
+
+from backend.app.generator.model import GenModel
+
+
+class CRUDGenModel(CRUDPlus[GenModel]):
+    pass
