+++ backend/app/generator/model/__init__.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 from backend.common.model import MappedBase  # noqa: I001
-from backend.app.generator.model.gen import GenModel, GenBusiness
+from backend.app.generator.model.gen_business import GenBusiness
+from backend.app.generator.model.gen_model import GenModel
