+++ backend/app/generator/schema/gen_business.py
@@ -35,28 +35,3 @@ class GetGenBusinessListDetails(GenBusinessSchemaBase):
     id: int
     created_time: datetime
     updated_time: datetime | None = None
-
-
-class GenModelSchemaBase(SchemaBase):
-    name: str
-    comment: str | None = None
-    type: str
-    default: str | None = None
-    length: int
-    is_pk: bool
-    is_nullable: bool
-
-
-class CreateGenModel(GenModelSchemaBase):
-    pass
-
-
-class UpdateGenModel(GenModelSchemaBase):
-    pass
-
-
-class GetGenModelListDetails(GenModelSchemaBase):
-    model_config = ConfigDict(from_attributes=True)
-
-    id: int
-    gen_business_id: int
