+++ backend/app/generator/service/gen_business_service.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+
+class GenBusinessService:
+    @staticmethod
+    async def create() -> None: ...
+
+    @staticmethod
+    async def update() -> int: ...
+
+    @staticmethod
+    async def delete() -> int: ...
+
+
+gen_business_service = GenBusinessService()
