+++ backend/app/generator/service/gen_model_service.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+
+class GenModelService:
+    @staticmethod
+    async def create() -> None: ...
+
+    @staticmethod
+    async def update() -> int: ...
+
+    @staticmethod
+    async def delete() -> int: ...
+
+
+gen_model_service = GenModelService()
