+++ backend/app/generator/service/gen_service.py
@@ -10,24 +10,6 @@ class GenService:
     async def get() -> dict: ...
 
     @staticmethod
-    async def create_business() -> None: ...
-
-    @staticmethod
-    async def update_business() -> int: ...
-
-    @staticmethod
-    async def delete_business() -> int: ...
-
-    @staticmethod
-    async def create_model() -> None: ...
-
-    @staticmethod
-    async def update_model() -> int: ...
-
-    @staticmethod
-    async def delete_model() -> int: ...
-
-    @staticmethod
     async def preview() -> dict: ...
 
     @staticmethod
